+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Fix sending test notification to a group integration
 - Fix the Login form to not perform form validation in GET requests
 - Fix special character escaping in ntfy notifications
+- Fix "Edit ntfy integration" page to fill the existing token in the form
 
 ## v3.0 - 2023-10-16